requirements.txt
discord==1.0.1
discord.py==1.6.0
python-dotenv==0.15.0
dnspython==1.16.0
youtube-dl==2021.2.10
PyNaCl==1.4.0